fix(clickpipes): add `stopped` attribute as a remediation to stuck Terraform plan
Current implementation uses `state` attribute in two-way interaction: for status feedback and control on ClickPipe lifecycle by stopping/starting ingestion.

This does not work well in Terraform. Produced state must be equal to plan, meaning, API could return different ClickPipe state, but Terraform would error with unexpected state value for `state` attribute.
This was remediated by introducing an extra condition in `syncClickPipeState` disallowing to persist back to the state if received value was `Failed` or `InternalError`. This was a workaround to not get Terraform complain on provider implementation.
In fact, it was a wrong logic, and it has only hidden a root cause of problem.

This PR splits status feedback and lifecycle control into two separate attributes:
- `state` attribute will be changed to a read-only reporting an actual state back. **This is considered as a breaking change, and will require user to remove `state` from theirs Terraform plan.**
- `stopped` attribute will allow control on pipe state. **If you had your `state = Stopped` previously, you will have to set `stopped = true`.**

This change is necessary to fix a control over pipe's configuration regardless of a current pipe state. `clickhouse_clickpipe` resource is in alpha builds only. It's considered safe to proceed with this change.

Additionally, post-apply status confirmation logic is fixed to retry backoff wait max to 120s.